First version of working PDF converter with wkhtml2pdf linkage
