--- .gitmodules
@@ -1,3 +1,0 @@
-[submodule "InvoiceLib/externals/wkhtmltopdf"]
-	path = InvoiceLib/externals/wkhtmltopdf
-	url = git@github.com:MickaelOnTheWave/wkhtmltopdf.git
